pyproject.toml
[tool.poetry]
name = "twitterproject"
version = "0.1.0"
description = ""
authors = ["Sezimar Jr <sezimar_jrcs@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
djangorestframework = "^3.15.2"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"
faker = "^33.0.0"
django = "^5.1.3"
pillow = "^11.0.0"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
